backend/requirements.txt
fastapi
uvicorn
pydantic
google-genai
pytest
httpx
python-dotenv
python-jose[cryptography]
